ventas y reportes

+++ Modelos/MReporte.cs
@@ -9,13 +9,13 @@ namespace proyectofinalwebII
     {
         public string Tipo { get; set; }
         public int Cantidad { get; set; }
-        public int Total { get; set; }
+        public Double Total { get; set; }
 
         public MReporte()
         {
         }
 
-        public MReporte(string tipo, int cantidad, int total)
+        public MReporte(string tipo, int cantidad, double total)
         {
             Tipo = tipo;
             Cantidad = cantidad;
